requirements.txt
streamlit==1.38.0
torch==2.4.0
torchvision==0.19.0
transformers==4.45.0
pillow==10.4.0
faiss-cpu==1.8.0.post1
numpy==2.0.0
matplotlib==3.9.0
pandas==2.2.0